+++ framework/include/fvkernels/FVAdvection.h
@@ -3,11 +3,11 @@
 #include "FVKernel.h"
 
 template <ComputeStage compute_stage>
-class FVConstantScalarAdvection : public FVFluxKernel<compute_stage>
+class FVAdvection : public FVFluxKernel<compute_stage>
 {
 public:
   static InputParameters validParams();
-  FVConstantScalarAdvection(const InputParameters & params);
+  FVAdvection(const InputParameters & params);
 
 protected:
   virtual ADReal computeQpResidual() override;
