+++ test/tests/kernels/fv_constant_scalar_advection/2D_constant_scalar_advection.i
@@ -33,7 +33,7 @@
 
 [FVKernels]
   [./advection]
-    type = FVConstantScalarAdvection
+    type = FVAdvection
     variable = v
     velocity = '1 0.5 0'
   [../]
